mapView: Use custom widget to embed ChamplainView
We'll be needing to overlay a sidebar and a button with transparent
container for (un)revealing it. Trouble with GtkChamplain.Embed is that
its a native window and using Gtk.Overlay with (semi-)transparent widgets
over native windows is not supported[1].

We now create our own embedding widget for ChamplainView that has its own
actor with BinLayout put ChamplainView in it. Later we simply add overlay
widgets using GtkClutter into it.

co-author: Zeeshan Ali (Khattak) <zeeshanak@gnome.org>

[1]
 https://git.gnome.org/browse/gtk+/commit/?h=gtk-3-8&id=8c02e290c443f7f99d8b26855cef17ab66a5bad6