Ensure current & dest. location visible before go-to
Lets first ensure that both current and destination location are
visible before we start the animated journey towards destination itself.
We do this to create the zoom-out-then-zoom-in effect that many map
implementations do. This not only makes the go-to animation look a lot
better visually but also give user a good idea of where the destination
is compared to current location.