MapView: Adapt to new geocode-glib's geocoding API
The geocoding API now returns Place objects with possibly a lot more
details, rather than simple Location objects.

For now we just take the Location out of the Place object and use it
like we have been doing so far but later we'll want to actually utilize
the information provided by Place.