MainWindow: move to CSD
Move maps to be client side decorated (CSD), that is: Gtk+ will draw the
window decorations instead of relying on the window manager for drawing
them.

This move has been made by a few other apps in GNOME already (like
devhelp) and is mandatory for making us work in the wayland tech demo
that will be released in 3.10.

This is done by just moving the GdHeaderBar inside a special child to
window with the attribute type = titlebar and that will be your new title
bar. I also added a close button like the other apps do, which will help
people close the app even when maximised.

https://bugzilla.gnome.org/show_bug.cgi?id=705537