MainWindow: escape markup for place description
It is possible to get search results that include characters like '&'.
We need to escape the markup sent to the search result treeview.

Otherwise we end up with warnings in the console like:
"(gnome-maps:2368): Gtk-WARNING **: Failed to set text from markup due to
error parsing markup: Error on line 1: Entity did not end with a
semicolon; most likely you used an ampersand character without intending
to start an entity - escape ampersand as &amp;"

https://bugzilla.gnome.org/show_bug.cgi?id=706653