Add a context menu
Currently this menu has only one item: "What's here?". If user activates
this item, a reverse geocoding is performed for the coordinates where
user right clicked and place information (currently only the name) is
shown in bubble (just like for forward geocoding search results).

Next would be to add a "I'm here" menu item to provide user the ability
to correct (or improve the accuracy) of his/her location.

https://bugzilla.gnome.org/show_bug.cgi?id=706726