Geoclue: Allow overriding location
Currently this only updates the concept of user location within Maps.

I'm hoping that later we can build something more sophisticed on top of
this: Allowing user to easily submit WiFi/IP information to server to
not only improve geolocation for the user submitting the data but also
other users (when) in the same area. This will be very important since
the open/free geolocation services we (will) use in geoclue don't have
enough coverage/accuracy.

Perhaps geoclue can provide an API for data submission:

https://bugs.freedesktop.org/show_bug.cgi?id=68536

https://bugzilla.gnome.org/show_bug.cgi?id=706726