mapLocation: set saner goto animation duration
Champlain calculates the goto_animation_duration using zoom_level.
It also treats ensure_visible as a go_to with its own duration time.

In Maps, each goTo is a combination of a call to Champlain's
ensure_visible and go_to.

This patch calculates the goto animation duration using the distance
between the two locations. It also divides the calculated value by
two to get the duration for our whole trip that includes a call
to ensure_visible.

https://bugzilla.gnome.org/show_bug.cgi?id=722263