searchPopup: make GtkStack the top-level widget
It is bad practice to add children that can not scroll to a
GtkScrolledWindow.

This patch makes the GtkStack the top-level widget of the
search-popup ui file. With this patch the GtkTreeView will be the
only child of the GtkScrolledWindow.

This change also fixes a buglet where the spinner would not be
centered in the visible popup when scrollbars where visible.

https://bugzilla.gnome.org/show_bug.cgi?id=722869