MainWindow: Make clicking the map grab focus
If the map does not grab focus on click it will be impossible to
dismiss the on-screen keyboard it you are using a touch/tablet device.

Set can-focus to true for the GtkOverlay where the map is contained.
Have the map grab focus after clicking on it or after selecting an
item in the SearchPopup.

https://bugzilla.gnome.org/show_bug.cgi?id=709604