notificationManager: Simplify showNotification
Right now showNotification takes a tuple of name and class to create
and cache a kind of simpleton Notification class. This makes the API
a bit awkward.

Instead we can just have the method take a Notification instance
directly. If the caller want the functionality of re-usable
notifications they can do the simpleton creation on their end.
The method still checks if the notification instance is already
in the overlay.

https://bugzilla.gnome.org/show_bug.cgi?id=727706