placeEntry: Return true from 'match-selected'
The documentation for GtkEntryCompletion states:

"When the user selects a completion, the content of the
entry is updated. By default, the content of the entry is
replaced by the text column of the model, but this can be
overridden by connecting to the "match-selected" signal and
updating the entry in the signal handler. Note that you
should return TRUE from the signal handler to suppress the
default behaviour."

We update the entry in our signal handler but we do not
return TRUE. This results in the text being set to the
empty string and setting the place of the placeEntry
to null.

https://bugzilla.gnome.org/show_bug.cgi?id=734597