Cleanup: Remove type coercion compares
The '==' and '!=' operators in JS is a bit weird since it tries to do
automatic and sometimes unexpected type coercions instead of a strict
comparison.

In Maps we have decided to use the '===' and '!==' operators instead
since they do stricter checks.

This patch converts two cases where '!=' passed through review.

https://bugzilla.gnome.org/show_bug.cgi?id=699967