Place: translate opening hours
Parse the opening_hours tag from OSM to produce translatable
strings.

To make it manageable, handle up to three different time specifications
(those separated by ;).

Each day range can contain up to three comma-separated day lists of each
either a single day or in the form start-end.

For each component, the time specification list can contain one or two
elements (each with start and end time, typically two are used i.e. when
there is a lunch break).

This is done to avoid doing a list iteration which could potentially be
difficult to translate right in some cases.

Furthermore, 24/7 and 00:00-24:00 are recognized as "around the clock".
It also handles the special cases PH and SH for public and school holidays.

https://bugzilla.gnome.org/show_bug.cgi?id=740836