main: Add required GIR dependencies
Use pkg.require to mark a set of dependencies on GI and standard
JS modules. The keys are repository names and the values are API
versions.

This will not specify the required _version_ of the package we
depend on. But it will make sure that the required package is
installed.

If a check fails a message will be printed and the application will
quit.

https://bugzilla.gnome.org/show_bug.cgi?id=742380