Start with last viewed location
When users open Maps they start with map of the world. This is less
convenient than starting at where they were last time. So through this
patch we store their last visited location in gsettings and retrieve
it back next time they open.

Whenever there is a view-moved event we schedule a timeout to save the
current location. When a new timeout is sceduled, we check if we have
a pending timeout, if we do, we dont add the new one. This ensures that
we are not saving locations too many times.

https://bugzilla.gnome.org/show_bug.cgi?id=755847