Add a help overlay
GNOME has a new Goal: Adding shortcut windows.

GTK+ 3.19.x includes a new widget called GtkShortcutsWindow.
This is a dialog window that shows an overview of shortcuts
(keyboard shortcuts and touch gestures) for an application.

To add a help overlay we need to create an UI file in our
resources at $(gresource prefix)/gtk/.

And the rest should be magical.

https://bugzilla.gnome.org/show_bug.cgi?id=756946