GeoJSON/Layers: Formalize and add GUI
Maps supports displaying GeoJSON files as overlays on top of the map via
drag and drop or by specifying the file on the command line.

This commit adds the capability to select and load these files via the
UI via a button in the Layers Popover; furthermore, the Layers Popover
displays which overlays are loaded and allows the user to remove them at
will.

https://bugzilla.gnome.org/show_bug.cgi?id=759544