mainWindow: set correct DND drag status
Sometimes Maps wouldn't accept a drop because we weren't setting the
drag_status appropriately.

Also change drag_finish to denote that the originally-dropped file is
not a semantic move and that the source shouldn't delete the original
data because Maps doesn't really take responsibility for the file -- it
just displays it until the app quits.

https://bugzilla.gnome.org/show_bug.cgi?id=759544

https://bugzilla.gnome.org/show_bug.cgi?id=760256