mapWalker: Improve travel animations
Reorder the animation callbacks. There's some sort of race
condition where the callbacks trigger sometimes for different
animations than the one you expect. This leads to things like the map
zooming-in instead of out to pan, and the map zooming-in on an area
different than the ending position. Reordering the expressions seems
to make us win these races.

The ChamplainView API doesn't allow us to do much better unfortunately
so this is a quick bandaid until there's time for bigger changes to be
made.

https://bugzilla.gnome.org/show_bug.cgi?id=760256