Implement ChamplainExportable interface for markers
This cycle we added a way to export the ChamplainView
to a cairo surface. This is meant as way of creating PNG
images of our map. As well as a way to facilitate printing
of the map view.

The cairo_view_to_surface function will go through each
ChamplainLayer added to the view and check if it implements
ChamplainExportable, and add it to the final surface if
it does. The ChamplainMarkerLayer will only export surface
if the markers within implements ChamplainExportable.
Let's make it so.

https://bugzilla.gnome.org/show_bug.cgi?id=760352