css: Replace deprecated properties
The problem is that these two warning pop up everytime we try
to run gnome-maps due to GtkWidget deprecation (since version 3.14).

(org.gnome.Maps:13949): Gtk-WARNING **: Theme parsing error:
application.css:13:33: The style property GtkWidget:focus-line-width
is deprecated and shouldn't be used anymore. It will be removed in a
future version

(org.gnome.Maps:13949): Gtk-WARNING **: Theme parsing error:
application.css:14:30: The style property GtkWidget:focus-padding is
deprecated and shouldn't be used anymore. It will be removed in a
future version

To fix this, we remove 'GtkWidget-focus-line-width' and
'GtkWidget-focus-padding' and use 'outline-width' and 'padding' insted,
which is a CSS standard property and a suitable replacement.
This does give the same behavior.

https://bugzilla.gnome.org/show_bug.cgi?id=760444