osmEdit: Use the call proxy when checking the OAuth access token
We can't rely on the sign-in proxy being availble to check for the
OAuth access token when trying to upload a changeset, since the
changes in https://bugzilla.gnome.org/show_bug.cgi?id=759464
since that proxy object will only be available right after a sign-in
operation now.
Instead check the call proxy, which is also updated on successfull
sign-in and in the keyring lookup callback.

https://bugzilla.gnome.org/show_bug.cgi?id=761019