Introduce an abstract base class for search result popovers
The base class for search popover enables propagating typing events
to an associated search entry. This allows the user to continue
typing when the search result popover is activated (while still being
able to navigate the search results in the popover and selecting
a row with enter).
This code was earlier part of the place search popover (in the main
header bar).
Also rename the place search popover from SearchPopup to
PlacePopover to be consistent with the search result popover for
OSM POI results (in a later patch).

https://bugzilla.gnome.org/show_bug.cgi?id=761327