Lint: Fix leaking vars in case stmts
The ESLint recommended settings disallows declaring new variables with
let and const in case-statements. This is because a case statement
doesn't produce a new scope, which might be confusing and be a source of
bugs.

When there is need to use variables inside a switch-statement it is just
as easy to declare it outside or wrap the case-clause inside {}'s.