osmEdit: Add support for editing phone number
Adds the ability to edit phone numbers. Also reformat URIs using
the tel: scheme by stripping off the leading tel: and trailing
parameters (following a ”;”), which could be useful when copying
links i.e. from a browser.
Also add phone number to the information shown in the expanded area
in the place bubble.

https://bugzilla.gnome.org/show_bug.cgi?id=761636