placeBubble: Use label for links
Since 2.18, GTK+ supports markup for clickable hyperlinks
in addition to regular Pango markup.

The markup for links is borrowed from HTML, using the <a>
with “href“ and “title“ attributes. GTK+ renders links
similar to the way they appear in web browsers, with colored,
underlined text.

The “title“ attribute is displayed as a tooltip on the link.