osmUtils: Don't update a place's name if not in the OSM object
When a place has been edited on OSM don't reset the Place's name attribute
if it's not set in the OSM object, since the PlaceFormatter assumes this
is set. For bare address nodes, this would be set to a value contructed
from the address components, so we shouldn't remove this just because
the OSM object doesn't have a name set.
This fixes stack trace when saving places like that.

https://bugzilla.gnome.org/show_bug.cgi?id=762569