Routing: Remove latest place on failure
If a new entry is added to exixting route which results in
route not found then revert back to previous route and set the place
of the point to null. To implement this Add a latest field to routeQuery
which keeps the latest point added. and in routeService if route
is not found set latest.place to null.

https://bugzilla.gnome.org/show_bug.cgi?id=762483