mapView: Add a view property to indicate showing a route
Adds a new boolean property that is set to true when an actual route
is rendered on the map.
This will allow only showing the print button when a transit itinerary
is "dived into".
Also rename the current "routeVisible" property to "routingOpen" to
avoid confusion, since the new property is what's set when a route is
actually rendered on the map, the former property is set when routing
is ongoing (and the routing headerbar button is pressed).

https://bugzilla.gnome.org/show_bug.cgi?id=755808