flatpak: add org.gnome.Maps.json manifest
This manifest allows Maps to easily be built with all of its dependencies
inside of GNOME Builder. Just clone the project, select the runtime in
Build Preferences, and Build/Run to generate a flatpak and run.

Maps maintainers: feel free to revert if you don't want this, but I do
think it will simplify your "newcomer" process.

Requires Builder 3.22.3 (package or flatpak) and flatpak-builder on
the host system.