contextMenu: Avoid double declaration with 'let'
The following is a syntax error in ES6:

    let a = 'something';
    let a = 'other thing';

Previously GJS would silently accept this, but in GJS 1.48.0 it will be
a syntax error. This fixes the one instance where it happens. In this
case the offending variable was a dummy variable in a destructuring
assignment, so we can just elide the variable.

https://bugzilla.gnome.org/show_bug.cgi?id=778645