xmldom: Remove __proto__ accessor
This is based on a change in upstream change. I kinda gave up on importing
the new upstream files, as we did some patching previously, and just
using the verbatim upstream files gives errors and warnings.
At least, with this change loading GPX and KML layers still seem to work
fine.

https://bugzilla.gnome.org/show_bug.cgi?id=779867