modified files: 1)context-menu.ui 2)contextMenu.js 3)favoriteEditDialog.js(New File) 4)zoominFavoriteDialog.js(New File) 5)org.gnome.Maps.src.gresource.xml 6)zoom-in-favorite-dialog.ui 7)favorite-edit-dialog.ui 8)org.gnome.Maps.data.gresource.xml
New Feature Added:
Issue #144
Not all streets/roads have their street numbers defined in OpenStreetMap so, on a back road in the middle of nowhere, you can't create a favorite at an exact location. I suggest to add a "Create favorite" in the context menu when we right-click on a specific point on the map, which would pick the exact GPS coordinate of the point where we right click and create a favorite based on these coordinates.

Overview of the fix:
The fix is designed using the OSM edit dialog files as reference.

https://gitlab.gnome.org/GNOME/gnome-maps/issues/144