mapView: Do not animate to stored location
This avoids the gotoBBox function, which uses animations, and instead
sets the zoom level and calls center_on() on the view.

This involves storing the last viewed location by coordinates and zoom
level rather than by a bounding box. A migration path is provided so
that the old bounding boxes use the old, animated method.