routeQuery: Require all filled points for valid
Only consider query as valid if all points are filled.
This avoids a case where adding a new destination using
the context menu results in two requests being executed,
first one with the previous places and immediatly one
with the new point set to a place. This results in
unnessesary API calls and a glitchy UX.

Fixes #224