sendToDialog: Ensure consistency when marker moves
The Send To dialog now remembers the exact coordinates, not just the
Place object, from when it was created. This way, if the place moves
(such as a moving current location marker), the coordinates shown in the
dialog remain the same as when the Send To button was pressed.

Also, the map bubble now creates a new sendToDialog every time the Send
To button is pressed, so that the latest marker position is shown in the
dialog.