build: Fix libgfbgraph dependency
GFBGraph has been updated to 0.3, which caused Maps to fail with the error
"Unsatisfied dependency: GFBGraph" if built with freshly downloaded
dependencies. Fixed by updating the package requirement in main.js.

Also added the dependency in Meson. This way, if it happens again in the
future, we'll get a build error rather than a runtime error.