mainWindow: Adaptive routing sidebar
Makes the routing sidebar adaptive using HdyLeaflet. When the routing sidebar
is opened, if the window is too narrow, the map view will be hidden and only
the sidebar will be shown. The headerbar will change to have a back button,
which closes the sidebar.

If the window is wide enough to fit both map and sidebar, nothing happens
differently than before, except that the action bar no longer extends
underneath the sidebar.

Added libhandy as a dependency in order to use HdyLeaflet and HdyTitleBar.

Fixes #264.