sidebar: Only focus on next route entry when empty
Only automatically set focus on the next route entry
when selecting a place and the next one is not set
with a place (is empty).
This fixes a race-condition-like bug when dragging
markers around (setting raw coordinates) in some
cases. It also make more sense only to autofocus
on an empty entry leading the user into searching
for a "to" place.

Fixes #329