sidebar: Grey-out transit switch when not available
Due to a bug in GTK where setting visible false on a linked
button not working properly, we used to call destroy()
on this button when transit providers where not available.
Since calling destroy() on widgets will not be supported
in GTK 4 and we could instead dim the button in this case
as this feature has now been exposed by default for a while.
And originally it was hidden since the feature where itself
hidden.