transitPlan: Round up duration to nearest minute
Round up duration value for display to the nearest
rounded up integer number of minutes, as displayed
departure time will be truncated down to the nearest
minute. This way the total number of minutes will always
match up instead of getting a rounding off-by-one
discrepancy in the case where the first leg has a
"partial" minute, and would be truncated down to the
next lower minute when showing the departure time.