graphHopper: Don't fold keep left/right instructions
In #171 we introduced "folding" of routing instructions
for "continue straight" and keep left or right instructions
to work around GraphHopper giving a lot of extra instructions
while staying on the same road number/name.
This has tended to take away a bit _too_ much, such as
"keep right" instructions when exiting a highway (and the
offramp has the same ref number as the main highway).
So, only fold exact "continue on street" instructions now.