Use external browser to authorize OAuth token
Drop the "fully automatic" sign in method that
we had adapted from JOSM. Instead implemtent
OAuth setup redirecting to an external web
browser. This is the preferred way to do
OAuth sign up, and also lets us drop the
dependency on WebKitGTK, as we don't need
to render the verification web view.