Migrate to ES6 modules
Switch to using ES6 modules for internal
modules and GI modules instead of using
the legacy imports object.
The unit tests are still using imports.jsunit
as this is not available as an ES6 module
(the preferred solution now is to use
jasmine-gjs). imports.bytearray,
imports.format, and imports.mainloop are
covered in follow-up commits.