general: Rework context menu
The context menu had a few papercuts after the port, notably
placement. The code was also a bit hard to follow - the actions
were declared in mainWindow.js, but defined in contextMenu.js.

This commit removes the ContextMenu class, and instead builds
and displays the popover in mapView.js. This also allows
us to move our actions to mapView.js using a custom action group.
The placement is also fixed.