placeMarker: Rename the mark-location icon
This icon is present in a symbolic recolorable
variant in some system theme packages.
This can result in a hard-to-read light grey
icon when dark theme is used, which gives
bad contrast against the (mostly light) raster
tiles. So rename the bundled icon with a maps-
prefix to prevent this.