file-data-source: Check for error on file retrieval
On local data sources, sometimes file read operations are cancelled.
When such a thing arises, empty data is passed to "received-data"
signal, leading to a warning in shumate: "Failed to create texture
from tile data: Unrecognized image file format", and a crash of
gnome-maps. We add a check on the result of
g_file_load_contents_finish and if there is an error we emit a
warning and we do not send the "received-data" signal.