mainWindow: Remove shortcuts for using bare +,=, and - for zooming
Supporting these (without modifier) causes issues when trying to type
in e.g. a place name containing a - in a place name in the search entry.
We used to have some workaround in the GTK 3 days, overriding the key-pressed
signal and filtering depending on the widget the event originated from to
support typing these characters into a search entry without triggering the
shortcut.
But this is not working anymore, and it seems to not be worth the effort
of trying to actually support this.
So we're probably better off dropping these. Can still support num pad
"add" and "subtract" keys still for zooming.