place: Store osmTags
Instead of storing each property individually, store all of the OSM
tags in an object. Then, in each property's getter, we return the value
from the tags.

There is no migration path from the old properties to the new ones,
except for a few that are critical for showing addresses. We simply
mark any place with no osmTags as stale and fetch them fresh from the
server.