overpass: Don't create place through Photon
Previously, when the Overpass code created a new place (when you enter
an openstreetmap.org URL in the search bar), it converted the results to
Photon format, then used the Photon code to create the place. Now we
create the place directly, using OSM tags.

Also in this change, we compute osmKey from osmTags when available.