placeStore: Migrate to new file
Changed the format of the place store file, changed its location so that
old versions of Maps don't try to read the new version, and added code
to migrate from the old version to the new.

- The old format had a JSON array at the root. The array has been moved
to the "places" key of a root object so that other properties can be
added.
- Added migration and version properties to the root object.
- Removed the language property. It will no longer be needed after we
store all OSM tags.
- StoredRoute paths are now stored in Encoded Polyline format, which is
far more compact.
- Renamed added to updated.
- Added viewOrd instead of sorting the list by added date.
- Use camelCase for all stored properties
- The notion of favorites has been separated from the type property into
an isFavorite property. The type property now refers only to which Place
subclass to use. Also, the type property is moved to the Place object.
- The migration deletes Contact places, which have not been handled for
a while.

Also, the place store file is now pretty-printed when in developer mode.