mainWindow: Don't focus on the search enty on startup
This is not needed now that the map is set as key capture
for the main search entry.
Also adjust the click controller to correctly dismiss
the search popover on background clicking after previous
change in click handling the map view overlay.
Also grab focus on map when closing the sidebar.