overpass: Decrease POI deduplication distance
The maximum distance at which nearby POIs
with the same (or unnamed) name is ignored
to avoid nearby duplicates (such as clusters of
things like post boxes) appearing as individual
POIs with the same distance was set a bit high.
Increase this to avoid ignoring instances that
are not actually part of the same location.