placeBar: Fix place bar not appearing
There was a bug, I think since the port to GTK 4, where the place bar
would not appear. The click gesture on the marker was not claimed, so it
bubbled up to the map, and clicking the map immediately hides the place
bar again. Fixed by claiming the click on the marker.

Also, changed the gesture that hides the place bar so that it doesn't
trigger when the map is panned.