mapStyle: Include labels for islands
Also introduce a way for place definitions
to defined a max rank for filtering so
that smaller islands (higher rank) can
be filtered out at lower zoom levels.
And add support for an optional id
to override the autogenerated one
based on classes to allow multiple
place definitions using the same classes
but different parameters, such as maxRank.
This is needed for compatibility with MapLibre.