place: Derive country code for coordinates as fallback
Use coordinates to determine country if not set as a
construct parameters, or available via the addr:country
tag. This fixes getting the right address format when
clicking directly on a POI icon on the map when there's
no addr:country tag set, getting the same behavior as
when instanciating the place from a search result
(via the Photon parser).