printLayout: Remove code to draw minimaps
This was never fully ported to work with
libshumate and has been commented-out.
Remove it for now, as it depends on
the "print source" from the service
defintion that was connected to a
raster rendererer. So remove it
for now, and possibly re-implement it
later on building on vector source.