placeView: Show place icon and type
Show the place icon and name, when we have a mapping
for it in the definitions except for places of category
"place", as it would be a bit superflouse for places
like cities and towns. For unnamed places, show the
place type along with its icon in place of the name
(in bold).