transitPathLayer: Ensure contrasting outline color
Safe-guard with a contrasting outline color.
This corresponds with how it's done for
the boarding markers and avoids using
the default color in cases where
there's only a route color, but no text color
defined. As otherwise the default text color
would be chosen in cases where the color is
low contrast relative to the backgroud (light in
light mode, or dark in dark mode).