mapView: Always set selected place directly
Don't wait for animation to show selected place.
Since we now show the selected place in the
sidebar or bottom sheet rather than a popover
attached to the marker, it makes sense to show
it directly. It also fixes the issue with the
animation finished signal not being emitted
when animations are disabled in GTK.